Split code into new structure
- moved code in /chesspi/api.py to the following:

 - runserver.py --> run the app
 - chesspi/__init__.py --> initialize app and routes
 - chesspi/resources/games.py --> /games logic
 - chesspi/resources/index.py --> / logic
