+++ chesspi/__init__.py
@@ -1,0 +1,14 @@
+# Python v3.5.1
+
+from flask import Flask
+from flask_restful import Api
+from chesspi.resources.games import Game, GameList
+from chesspi.resources.index import Index
+
+app = Flask(__name__)
+api = Api(app)
+
+# API Routing
+api.add_resource(Index, '/')
+api.add_resource(Game, '/games/<game_id>')
+api.add_resource(GameList, '/games')
