+++ runserver.py
@@ -1,0 +1,2 @@
+from chesspi import app
+app.run(debug=True)
